Updated repository package name to the agreed upon package name
