+++ pom.xml
@@ -2,7 +2,7 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
-    <groupId>com.softmed</groupId>
+    <groupId>tz.go.moh.him</groupId>
     <artifactId>hdr-mediator-emr</artifactId>
     <version>0.1.0</version>
     <properties>
@@ -41,7 +41,7 @@
                             </descriptorRefs>
                             <archive>
                                 <manifest>
-                                    <mainClass>com.softmed.hdr_mediator_emr.MediatorMain</mainClass>
+                                    <mainClass>tz.go.moh.him.hdr.mediator.emr.MediatorMain</mainClass>
                                 </manifest>
                             </archive>
                         </configuration>
