+++ src/main/java/tz/go/moh/him/hdr/mediator/emr/Constants.java
@@ -1,0 +1,14 @@
+package tz.go.moh.him.hdr.mediator.emr;
+
+public class Constants {
+
+    public interface ErrorMessages {
+        String ERROR_REQUIRED_FIELDS_CHECK_FAILED = " - Required fields are empty;";
+        String ERROR_INVALID_PAYLOAD = "Invalid payload;";
+        String ERROR_SERVICE_DATE_IS_OF_INVALID_FORMAT_IS_NOT_A_VALID_PAST_DATE = " - service date is of invalid format/is not a valid past date;";
+        String ERROR_TRANSACTION_DATE_IS_OF_INVALID_FORMAT_IS_NOT_A_VALID_PAST_DATE = " - transaction date is of invalid format/is not a valid past date;";
+        String ERROR_DATE_DEATH_OCCURRED_IS_OF_INVALID_FORMAT_IS_NOT_A_VALID_PAST_DATE = " - date death occurred is of invalid format/is not a valid past date;";
+        String ERROR_ADMISSION_DATE_OCCURRED_IS_OF_INVALID_FORMAT_IS_NOT_A_VALID_PAST_DATE = " - admission date is of invalid format/is not a valid past date;";
+
+    }
+}
